tests/requirements-test.txt
# Test requirements for IBEX AI tests
pytest>=7.0.0
pytest-asyncio>=0.21.0
PyYAML>=6.0
aiohttp>=3.8.0

# Optional for enhanced testing
pytest-cov>=4.0.0
pytest-mock>=3.10.0
